Special date i18n commit.
This mirrors the previous commit in that we use FMTE^XLFDT to do date
formatting for non-US locales; however, the Vitals Delphi code changed
the '-' to '/' in order for Windows in a US locale to be able to
properly parse a US date. That's fine and well for a US locale Windows
machine, but if we send back a Korean locale (which uses - as
delimiters) and change them to '/', the date will no longer parse in
Windows.

The solution is to send the proper Windows parsable dates from VistA (as
fragile as that may be; but that problem has to do with the design of
the Vitals package RPCs and Delphi Code). Delphi will no longer change
the '-' to '/'.